dom/base/test/test_bug685798.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=685798
-->
<head>
  <title>Test for Bug 685798</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=685798">Mozilla Bug 685798</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 685798 **/


is(document.parentElement, null,
   "Document shouldn't have parentElement.");
is(document.documentElement.parentElement, null,
   "DocumentElement shouldn't have parentElement.");
is(document.documentElement.firstChild.parentElement, document.documentElement,
   "DocumentElement's child should have DocumentElement as parent.");

var df = document.createRange().createContextualFragment("<div>foo</div>");
is(df.parentElement, null,
   "DocumentFragment should be null.");
is(df.firstChild.parentElement, null, 
   "DocumentFragment's child shouldn't have parentElement");
is(df.firstChild.firstChild.parentElement, df.firstChild,
   "Text node's parent should be element.");

is(document.createTextNode("foo").parentElement, null,
   "Text node shouldn't have parent element.");


</script>
</pre>
</body>
</html>
